requirements.txt
Bootstrap-Flask==1.2.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
Flask==1.1.1
Flask-Caching==1.8.0
Flask-Dance==3.0.0
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.1
httplib2==0.17.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
oauth2==1.9.0.post1
oauthlib==3.1.0
pkg-resources==0.0.0
python-dotenv==0.12.0
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
SQLAlchemy==1.3.15
SQLAlchemy-Utils==0.36.3
urllib3==1.25.8
URLObject==2.4.3
Werkzeug==1.0.0
